requirements.txt
duckdb>=0.10.0
pymongo>=4.6.0
python-dotenv>=1.0.0
